+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - registry/storage/cache/redis: Remove maxidle and maxactive pool settings
 - configuration: Drop support for logstash and combined log formats and default to json
 - configuration: Drop support for log hooks
+- configuration: Drop NewRelic reporting support
 
 ## [v2.13.0-gitlab] - 2020-12-15
 ### Added